#7
Ajustes CRUD produto, frontend.

+++ ezbuy-project-frontend/templates/ezbuy-app/src/app/screens/product/product.component.ts
@@ -6,6 +6,7 @@ import { ProductService } from '../../services/product.service';
 import { response } from 'express';
 import { ToastrService } from 'ngx-toastr';
 import { Observable, of, switchMap, take } from 'rxjs';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product',
@@ -15,16 +16,11 @@ import { Observable, of, switchMap, take } from 'rxjs';
   styleUrl: './product.component.scss'
 })
 export class ProductComponent implements OnInit {
-  // products = [
-  //   { id: 1, name: 'Produto 1', description: 'Descrição do produto 1', category: 'Categoria 1', price: 10.99 },
-  //   { id: 2, name: 'Produto 2', description: 'Descrição do produto 2', category: 'Categoria 2', price: 20.99 },
-  //   { id: 3, name: 'Produto 3', description: 'Descrição do produto 3', category: 'Categoria 1', price: 15.99 },
-  //   // Adicione mais produtos conforme necessário
-  // ];
-
   products: IProduct[] = [];
 
-  constructor(private router: Router, private productService: ProductService, private toastrService: ToastrService,){}
+  constructor(private router: Router, private productService: ProductService, private toastrService: ToastrService,){
+
+  }
 
   ngOnInit(){
     this.loadProducts();
@@ -52,14 +48,17 @@ export class ProductComponent implements OnInit {
 
   delete(product: any){
     if (product != null){
-      this.toastrService.warning('Tem certeza que deseja excluir este item?', 'Confirmação', {
-        closeButton: true,
-        timeOut: 0,
-        extendedTimeOut: 0,
-        onActivateTick: true,
-        disableTimeOut: true,
-      }).onTap.pipe(
-        switchMap(() => {
+      Swal.fire({
+        title: 'Tem certeza?',
+        text: 'Você quer realmente deletar este item?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Sim, deletar!',
+        cancelButtonText: 'Cancelar'
+      }).then((result) => {
+        if (result.isConfirmed){
           this.productService.delete(product.id).subscribe((response) => {
             if (response){
               this.toastrService.success('Entidade excluída com sucesso!', 'Sucesso!');
@@ -69,10 +68,9 @@ export class ProductComponent implements OnInit {
               this.toastrService.error('Falha na tentativa de excluir a entidades!', 'Error!');
             }
           });
-          return of(true);
-        }),
-        take(1)
-      ).subscribe();
+        }
+      });
+
     }
 
     return;
